test(cuda.core): add note about using multiple conftest modules (#2596)
* test(cuda.core): add note about using multiple conftest modules

* tests refactoring

* include references to pytest docs
