+++ cuda_core/pytest.ini
@@ -4,6 +4,7 @@
 
 [pytest]
 addopts = --showlocals --durations=20
+pythonpath = tests
 norecursedirs = cython
 markers =
     # Keep this authorship marker registry in sync across all pytest config roots.
