+++ cuda_core/tests/graph/test_graph_definition_errors.py
@@ -6,8 +6,8 @@
 import ctypes
 
 import pytest
-from conftest import xfail_on_graph_mempool_oom
 from helpers.graph_kernels import compile_common_kernels
+from helpers.memory import xfail_on_graph_mempool_oom
 from helpers.misc import try_create_condition
 
 from cuda.core import Device, LaunchConfig
