+++ cuda_core/tests/graph/test_graph_definition_integration.py
@@ -7,7 +7,7 @@ import ctypes
 
 import numpy as np
 import pytest
-from conftest import xfail_on_graph_mempool_oom
+from helpers.memory import xfail_on_graph_mempool_oom
 
 from cuda.core import Device, EventOptions, LaunchConfig, Program, ProgramOptions
 from cuda.core._utils.cuda_utils import driver, handle_return
