+++ cuda_core/tests/graph/test_graph_definition_lifetime.py
@@ -13,8 +13,8 @@ import time
 import weakref
 
 import pytest
-from conftest import xfail_on_graph_mempool_oom
 from helpers.graph_kernels import compile_common_kernels
+from helpers.memory import xfail_on_graph_mempool_oom
 from helpers.misc import try_create_condition
 
 from cuda_python_test_helpers import under_compute_sanitizer
