+++ cuda_core/tests/test_launcher.py
@@ -4,7 +4,7 @@
 import ctypes
 
 import helpers
-from cuda_python_test_helpers.marks import requires_module
+from cuda_python_test_helpers.marks import requires_module, skipif_need_cuda_headers
 from helpers.misc import StreamWrapper
 
 try:
@@ -13,7 +13,6 @@ except ImportError:
     cp = None
 import numpy as np
 import pytest
-from conftest import skipif_need_cuda_headers
 
 from cuda.core import (
     Device,
