+++ cuda_core/tests/test_managed_memory_warning.py
@@ -11,7 +11,8 @@ These tests only run on affected platforms (concurrent_managed_access is False).
 import warnings
 
 import pytest
-from conftest import create_managed_memory_resource_or_skip, xfail_if_mempool_oom
+from cuda_python_test_helpers.mempool import xfail_if_mempool_oom
+from helpers.memory import create_managed_memory_resource_or_skip
 
 import cuda.bindings
 from cuda.core import Device, ManagedMemoryResource, ManagedMemoryResourceOptions
