+++ cuda_core/tests/test_object_protocols.py
@@ -13,9 +13,9 @@ import re
 import weakref
 
 import pytest
-from conftest import xfail_on_graph_mempool_oom
 from helpers.constants import POOL_SIZE
 from helpers.graph_kernels import compile_common_kernels
+from helpers.memory import xfail_on_graph_mempool_oom
 from helpers.misc import try_create_condition
 
 from cuda.core import (
